final_pjt_back/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==4.0.1
Django==4.2.4
django-admin-rangefilter==0.10.0
django-allauth==0.54.0
django-cors-headers==4.2.0
django-environ==0.10.0
django-extensions==3.2.3
django-filter==23.2
django-rest-auth==0.9.5
djangorestframework==3.14.0
exceptiongroup==1.2.0
executing==2.0.1
Flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
importlib_metadata==7.1.0
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
multidict==6.0.5
oauthlib==3.2.2
openai==1.30.1
parso==0.8.3
pexpect==4.9.0
Pillow==9.5.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
stack-data==0.6.3
tqdm==4.66.4
traitlets==5.14.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.12
Werkzeug==3.0.3
yarl==1.9.4
zipp==3.18.2
